Moved Record.js from Utils to Presentation Layer
